verfiy that order has record at time limit

+++ database/migrations/2025_01_28_101812_remove_expire_at_and_defualt_contract_signed_columns_from_trader_orders_table.php
@@ -1,7 +1,11 @@
 <?php
 
+use App\Models\TraderOrder;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
@@ -13,9 +17,46 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::table('trader_orders', function (Blueprint $table) {
-            $table->dropColumn(['expire_at', 'default_contract_sign_time_limit']);
-        });
+        DB::beginTransaction();
+        try {
+            $traderOrders = TraderOrder::whereNotNull('expire_at')
+                ->whereNotNull('default_contract_sign_time_limit')
+                ->get();
+
+            foreach ($traderOrders as $order) {
+                if (! $order->timeLimits()->exists()) {
+                    throw new ModelNotFoundException("Missing TraderLimit for TraderOrder ID: {$order->id}");
+                }
+
+                $order->update([
+                    'expire_at' => null,
+                    'default_contract_sign_time_limit' => null,
+                ]);
+            }
+
+            $traderOrdersCount = TraderOrder::whereNotNull('expire_at')
+                ->whereNotNull('default_contract_sign_time_limit')
+                ->count();
+
+            if ($traderOrdersCount == 0) {
+                Schema::table('trader_orders', function (Blueprint $table) {
+                    $table->dropColumn(['expire_at', 'default_contract_sign_time_limit']);
+                });
+            }
+
+            DB::commit();
+        } catch (ModelNotFoundException $e) {
+            DB::Rollback();
+            Log::error('TraderLimit check failed: '.$e->getMessage());
+            throw $e;
+        } catch (\Exception $e) {
+            DB::Rollback();
+            Log::error('Unexpected error: '.$e->getMessage());
+        }
+
+        // Schema::table('trader_orders', function (Blueprint $table) {
+        //     $table->dropColumn(['expire_at', 'default_contract_sign_time_limit']);
+        // });
     }
 
     /**
